Create Project_Euler_21_Amicable_Numbers
I built a program to determine the sum of all amicable numbers under 10,000 in Java on 7/1/2016. I am not going to post the code because 
the site asks that code for the problems not be posted online. Couple of comments on the problem, my original version captured perfect
numbers in addition to amicable numbers.
amicable numbers is where f(a) = b and f(b) = a, where the function is the sum of all perfect divisors for the number in question.
For instance f(220) = 284 and f(284) = 220. Thus, 284 and 220 are an amicable pair and both are referred to as amicable numbers.
A perfect number is where f(a) = a, again where the function is the sum of all perfect divisors. For instance, the divisors of 6 are 1, 2,
and 3, which add up to 6, making 6 the smallest perfect number.
My code originally captured these perfect numbers as well. The difficulty was that I was not sure that I needed to exclude them based
on the problem definition.

Another thing left out of the problem definition is whether or not to include amicable numbers whose pair is greater than 10,000. 
In my code, I only included amicable numbers where both numbers of the pair were under 10,000 and got the correct answer.