Add files via upload
Recording builds a class for a song having artist, title, and length in seconds as elements. RecordingSort instantiates five records (either manually through information from the user or built in based on selection by user). RecordingSort then asks by what the user wants the records sorted and proceeds to sort and then display the records sorted ascending based on the parameter selected by the user. It uses the bubble sort method without contacting the inner loop. I have a series of if statements at the heart of the two for loops used for bubble sort as the only way that I could come up with to integrate selecting by user for sorting since it is sorting based on a characteristic of an object. I would like to see a better way to implement the selection by the user for what to use for sorting.