Create Permutation
I completed the problem on Project Euler for determining the 1 millionth permutation of 0123456789. A permutation is use of the digits in a different order but not duplicating any of the digits. I found an algorithm on the internet that describes how to achieve a permutation (not implemented in code but just written in pseudocode). 
For the impatient, we will start with the actual algorithm. Here it is. Suppose that P[1..n] is a permutation of 1 through n. We can construct the next permutation in lexicographic order by following these simple steps:

1.Find the largest x such that P[x]<P[x+1].
(If there is no such x, P is the last permutation.)
2.Find the largest y such that P[x]<P[y].
3.Swap P[x] and P[y].
4.Reverse P[x+1 .. n].