Create Fibonacci1000_digits
I developed a program in Java to determine the index of the first Fibonacci number that has 1000 digits. I used BigInteger for this problem as well. The first fibonacci number with 1000 digits is 1070066266382758936764980584457396885083683896632151665013235203375314520604694040621889147582489792657804694888177591957484336466672569959512996030461262748092482186144069433051234774442750273781753087579391666192149259186759553966422837148943113074699503439547001985432609723067290192870526447243726117715821825548491120525013201478612965931381792235559657452039506137551467837543229119602129934048260706175397706847068202895486902666185435124521900369480641357447470911707619766945691070098024393439617474103736912503231365532164773697023167755051595173518460579954919410967778373229665796581646513903488154256310184224190259846088000110186255550245493937113651657039447629584714548523425950428582425306083544435428212611008992863795048006894330309773217834864543113205765659868456288616808718693835297350643986297640660000723562917905207051164077614812491885830945940566688339109350944456576357666151619317753792891661581327159616877487983821820492520348473874384736771934512787029218636250627816.

Notice that all of the digits are significant. That was a struggle that I had in C++, I couldn't get the part of boost that enables large integers to work and built in variables didn't have sufficient capability for significant figures. At the time I was using a number with 150 digits which a double is large enough to hold but a double only has roughly 15 digits of precision. That being said, this 1000 digit fibonacci number would not have worked trying to use a double either.

However, use of BigInteger is somewhat quirky. You can't use normal arithmetic operators so getting used to syntax for use of BigIntegers is a learning curve itself. Also, BigInteger numbers are not at all compatible with regular integers (as a seasoned programmer would likely expect). With numbers in the range of regular integers it is possible to convert back and forth but in my experience it takes some trial and error to get this working the way that you want it to.