Add files via upload
I created a couple of number guessing games. in NumberGuess, the user selects a number and the computer attempts to guess what the number is. It asks for prompts if the number selected is higher or lower than the guess. In NumberGuess2, the program generates a random number and asks the user to attempt to guess what it is. The program provides information whether the guess is higher or lower than the random number.