Add files via upload
This is an example of using object oriented programming. In C++ it is possible to set defaults for parameters in method parameter list. I learned through this exercise that Java does not support this. Instead, overloading of the method is used to support the different style of definition. With overloading I was able to change how price is calculated for regular candles and scented candles with scented candles being a sub-class of candles and price being defined only in the parent class.