Add files via upload
The five files Person, CollegeEmployee, Faculty, Student, and CollegeArrays make up a project. Person is a superclass, CollegeEmployee and Student are subclasses of Person and Faculty is a subclass of CollegeEmployee. These files enabling input of first name, last name, address, etc. information for various persons allowing the user to choose what type of person to input information for. It then prints to the screen all of the information for the various persons. It also includes built in data population to enable population without going through the set up process to set up each individual.